+++ __init__.py
@@ -41,20 +41,22 @@ except NameError:
     from .scripts import (
         addon_preferences,
         consts,
-        func_apply_as_shapekey,
-        func_apply_modifiers,
-        func_apply_modifiers_with_shapekeys,
-        func_package_utils,
-        func_select_axis_from_point,
-        func_separate_lr_shapekey,
-        func_separate_lr_shapekey_all,
-        func_separate_shapekeys,
-        func_utils,
         link_with_AutoMerge,
         link_with_MizoresCustomExporter,
         menu_edit_mesh_context,
         menu_object_context,
         translations,
+    )
+    from BlenderAddon_ShapeKeysUtil.scripts.funcs import (
+        func_package_utils,
+        func_select_axis_from_point,
+        func_apply_modifiers_with_shapekeys,
+        func_separate_shapekeys,
+        func_separate_lr_shapekey_all,
+        func_utils,
+        func_separate_lr_shapekey,
+        func_apply_as_shapekey,
+        func_apply_modifiers,
     )
     from BlenderAddon_ShapeKeysUtil.scripts.ops import (
         op_apply_modifiers, op_separate_lr_shapekey,
