+++ scripts/ops/op_separate_lr_shapekey.py
@@ -18,7 +18,8 @@
 
 import bpy
 from bpy.props import BoolProperty
-from .. import consts, func_utils, func_separate_lr_shapekey
+from .. import consts
+from ..funcs import func_utils, func_separate_lr_shapekey
 
 
 class OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey(bpy.types.Operator):
